# gatewayservice/gateway-service.js
const express = require('express');
const axios = require('axios');
const cors = require('cors');
const promBundle = require('express-prom-bundle');
//libraries required for OpenAPI-Swagger
const swaggerUi = require('swagger-ui-express'); 
const fs = require("fs")
const YAML = require('yaml')

const app = express();
const port = 8000;

const historyServiceUrl = process.env.WIKI_SERVICE_URL || 'http://localhost:8005';
const wikiServiceUrl = process.env.WIKI_SERVICE_URL || 'http://localhost:8004';
const llmServiceUrl = process.env.LLM_SERVICE_URL || 'http://localhost:8003';
const authServiceUrl = process.env.AUTH_SERVICE_URL || 'http://localhost:8002';
const userServiceUrl = process.env.USER_SERVICE_URL || 'http://localhost:8001';

app.use(cors());
app.use(express.json());

//Prometheus configuration
const metricsMiddleware = promBundle({includeMethod: true});
app.use(metricsMiddleware);

// Health check endpoint
app.get('/health', (req, res) => {
  res.json({ status: 'OK' });
});

app.post('/login', async (req, res) => {
  try {
    // Forward the login request to the authentication service
    const authResponse = await axios.post(authServiceUrl+'/login', req.body);
    res.json(authResponse.data);
  } catch (error) {
    res.status(error.response.status).json({ error: error.response.data.error });
  }
});

app.post('/adduser', async (req, res) => {
  try {
    // Forward the add user request to the user service
    const userResponse = await axios.post(userServiceUrl+'/adduser', req.body);
    res.json(userResponse.data);
  } catch (error) {
    res.status(error.response.status).json({ error: error.response.data.error });
  }
});

app.post('/askllm', async (req, res) => {
  try {
    // Forward the add user request to the user service
    const llmResponse = await axios.post(llmServiceUrl+'/ask', req.body);
    res.json(llmResponse.data);
  } catch (error) {
    res.status(error.response.status).json({ error: error.response.data.error });
  }
});

app.post('/questions/:kind', async (req, res) => {
  try {
    const wikiResponse = await axios.post(wikiServiceUrl + '/questions/' + req.params.kind, req.body);
    res.json(wikiResponse.data);
  } catch (error) {
    res.status(error.response.status).json({error: error.response.data.error });
  }
});

app.post('/savegame', async (req, res) => {
  try {
    const historyResponse = await axios.post(historyServiceUrl + '/savegame', req.body);
    res.json(historyResponse.data);
  } catch (error) {
    res.status(error.response.status).json({error: error.response.data.error });
  }
});

app.get('/gethistory', async (req, res) => {
  try {
    const historyResponse = await axios.get(historyServiceUrl + '/gethistory');
    res.json(historyResponse.data);
  } catch (error) {
    res.status(error.response.status).json({error: error.response.data.error });
  }
});

app.get('/getquestions/:id', async (req, res) => {
  try {
    const historyResponse = await axios.get(historyServiceUrl + `/getquestions/${req.params.id}`);
    res.json(historyResponse.data);
  } catch (error) {
    res.status(error.response.status).json({error: error.response.data.error });
  }
});

// Read the OpenAPI YAML file synchronously
openapiPath='./openapi.yaml'
if (fs.existsSync(openapiPath)) {
  const file = fs.readFileSync(openapiPath, 'utf8');

  // Parse the YAML content into a JavaScript object representing the Swagger document
  const swaggerDocument = YAML.parse(file);

  // Serve the Swagger UI documentation at the '/api-doc' endpoint
  // This middleware serves the Swagger UI files and sets up the Swagger UI page
  // It takes the parsed Swagger document as input
  app.use('/api-doc', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
} else {
  console.log("Not configuring OpenAPI. Configuration file not present.")
}


// Start the gateway service
const server = app.listen(port, () => {
  console.log(`Gateway Service listening at http://localhost:${port}`);
});

module.exports = server